Move misc.doccer to _lib.doccer (#9652)
* MAINT: misc: deprecate misc.doccer

Move the doccer module to scipy._lib, leave a deprecation shim in misc.

* MAINT: import doccer from _scipy.lib
