+++ scipy/io/matlab/miobase.py
@@ -19,7 +19,7 @@ if sys.version_info[0] >= 3:
 else:
     byteord = ord
 
-from scipy.misc import doccer
+from scipy._lib import doccer
 
 from . import byteordercodes as boc
