+++ scipy/ndimage/_ni_docstrings.py
@@ -1,7 +1,7 @@
 """Docstring components common to several ndimage functions."""
 from __future__ import division, print_function, absolute_import
 
-from scipy.misc import doccer
+from scipy._lib import doccer
 
 __all__ = ['docfiller']
